third_party/blink/web_tests/platform/generic/editing/deleting/delete-character-003-expected.txt

a|<br>
PASS sample.innerHTML is "<br>"
PASS selection.type is "Caret"
PASS selection.focusNode is sample
PASS selection.focusOffset is 0


a<br>|<br>
PASS sample.innerHTML is "a<br>"
PASS selection.type is "Caret"
PASS selection.focusNode is sample.firstChild
PASS selection.focusOffset is 1


a<br><br>|<br>
PASS sample.innerHTML is "a<br><br>"
PASS selection.type is "Caret"
PASS selection.focusNode is sample
PASS selection.focusOffset is 2


a|<br><br>
PASS sample.innerHTML is "<br><br>"
PASS selection.type is "Caret"
PASS selection.focusNode is sample
PASS selection.focusOffset is 0


a<br>|<br><br>
PASS sample.innerHTML is "a<br><br>"
PASS selection.type is "Caret"
PASS selection.focusNode is sample.firstChild
PASS selection.focusOffset is 1


a<br><br>|<br><br>
PASS sample.innerHTML is "a<br><br><br>"
PASS selection.type is "Caret"
PASS selection.focusNode is sample
PASS selection.focusOffset is 2


a<br>|b
PASS sample.innerHTML is "ab"
PASS selection.type is "Caret"
PASS selection.focusNode is sample.firstChild
PASS selection.focusOffset is 1


a<br><br>|b
PASS sample.innerHTML is "a<br>b"
PASS selection.type is "Caret"
PASS selection.focusNode is sample.lastChild
PASS selection.focusOffset is 0


a<br><br><br>|b
PASS sample.innerHTML is "a<br><br>b"
PASS selection.type is "Caret"
PASS selection.focusNode is sample.lastChild
PASS selection.focusOffset is 0


a<br>b|
PASS sample.innerHTML is "a<br><br>"
PASS selection.type is "Caret"
PASS selection.focusNode is sample
PASS selection.focusOffset is 2


a<br><br>b|
PASS sample.innerHTML is "a<br><br><br>"
PASS selection.type is "Caret"
PASS selection.focusNode is sample
PASS selection.focusOffset is 3


a<br><br><br>b|
PASS sample.innerHTML is "a<br><br><br><br>"
PASS selection.type is "Caret"
PASS selection.focusNode is sample
PASS selection.focusOffset is 4


PASS successfullyParsed is true

TEST COMPLETE
